+++ src/console/Cargo.toml
@@ -18,8 +18,6 @@ serde.workspace = true
 serde_cbor.workspace = true
 serde_json.workspace = true
 ciborium.workspace = true
-sha2.workspace = true
-hex.workspace = true
 regex.workspace = true
 semver = "1.0.25"
 junobuild-shared = { path = "../libs/shared" }
