+++ src/console/src/storage/state/stable.rs
@@ -2,7 +2,6 @@ use crate::strategies_impls::cdn::CdnStable;
 use junobuild_cdn::proposals::ProposalId;
 use junobuild_cdn::storage::ProposalAssetKey;
 use junobuild_collections::types::core::CollectionKey;
-use junobuild_shared::types::core::Blob;
 use junobuild_storage::types::state::FullPath;
 use junobuild_storage::types::store::{Asset, AssetEncoding, BlobOrKey};
 
@@ -12,10 +11,6 @@ pub fn get_asset_stable(
     full_path: &FullPath,
 ) -> Option<Asset> {
     junobuild_cdn::storage::stable::get_asset(&CdnStable, proposal_id, collection, full_path)
-}
-
-pub fn get_content_chunks_stable(encoding: &AssetEncoding, chunk_index: usize) -> Option<Blob> {
-    junobuild_cdn::storage::stable::get_content_chunks(&CdnStable, encoding, chunk_index)
 }
 
 pub fn get_assets_stable(proposal_id: &ProposalId) -> Vec<(ProposalAssetKey, Asset)> {
