+++ src/libs/cdn/src/proposals/mod.rs
@@ -6,6 +6,8 @@ mod workflows;
 pub use state::stable;
 pub use state::types::*;
 pub use types::*;
+pub use workflows::commit::*;
 pub use workflows::errors;
 pub use workflows::init::*;
+pub use workflows::pre_commit_assets::*;
 pub use workflows::submit::*;
