+++ src/libs/cdn/src/proposals/workflows/mod.rs
@@ -1,4 +1,7 @@
 mod assert;
+pub mod commit;
 pub mod errors;
 pub mod init;
+pub mod pre_commit_assets;
+mod store;
 pub mod submit;
